+++ tests/FunctionalTest.php
@@ -3,7 +3,6 @@
 namespace Pontedilana\WeasyprintBundle\Tests;
 
 use PHPUnit\Framework\TestCase;
-use Pontedilana\PhpWeasyPrint\Image;
 use Pontedilana\PhpWeasyPrint\Pdf;
 use Symfony\Component\DependencyInjection\Exception\ServiceNotFoundException;
 use Symfony\Component\Filesystem\Filesystem;
@@ -40,11 +39,6 @@ class FunctionalTest extends TestCase
 
         $this->assertInstanceof(Pdf::class, $pdf);
         $this->assertEquals('weasyprint', $pdf->getBinary());
-
-        $this->assertFalse($container->has('weasyprint.image'), 'The image service is available.');
-
-        $this->expectException(ServiceNotFoundException::class);
-        $image = $container->get('weasyprint.image');
     }
 
     public function testChangeBinaries(): void
@@ -60,13 +54,6 @@ class FunctionalTest extends TestCase
 
         $this->assertInstanceof(Pdf::class, $pdf);
         $this->assertEquals('/custom/binary/for/weasyprint', $pdf->getBinary());
-
-        $this->assertTrue($container->has('weasyprint.image'));
-
-        $image = $container->get('weasyprint.image');
-
-        $this->assertInstanceof(Image::class, $image);
-        $this->assertEquals('/custom/binary/for/weasyprint', $image->getBinary());
     }
 
     public function testChangeTemporaryFolder(): void
@@ -89,16 +76,5 @@ class FunctionalTest extends TestCase
         $container = $this->kernel->getContainer();
 
         $this->assertFalse($container->has('weasyprint.pdf'), 'The pdf service is NOT available.');
-    }
-
-    public function testEnableImage(): void
-    {
-        $this->kernel->setConfigurationFilename(__DIR__ . '/fixtures/config/enable_image.yml');
-        $this->kernel->boot();
-
-        $container = $this->kernel->getContainer();
-
-        $this->assertTrue($container->has('weasyprint.pdf'), 'The pdf service is available.');
-        $this->assertTrue($container->has('weasyprint.image'), 'The image service is NOT available.');
     }
 }