+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 2.0.0 - 2023-01-16
+### Changed
+- Drop support for unsupported php-weasyprint versions
+- Bump Symfony 6 support to >= 6.2
+
+### Removed
+- Remove deprecate Image response
+
 ## 1.4.0 - 2023-01-16
 ### Added
 - Add support for php-weasyprint 0.13